chore(lint): fix ruff 0.16 findings so CI lint gate passes
The workflow installs unpinned ruff; rules added since the last master run
(I001, SIM1xx, BLE001, RUF046/100, UP012) now fail the lint job on every PR.
Auto-fixes for import sorting and the mechanical rules; the eight deliberate
blanket 'except Exception' fallback guards keep their behavior and are
annotated with noqa: BLE001 instead of being narrowed.

+++ custom_components/rinnai/sensor.py
@@ -21,6 +21,7 @@ from homeassistant.const import (
 )
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
+
 from . import RinnaiConfigEntry
 from .entity import RinnaiEntity
 
@@ -358,7 +359,7 @@ class RinnaiSensor(RinnaiEntity, SensorEntity):
         if isinstance(value, (int, float)):
             adjusted_value = value * self.entity_description.value_multiplier
             if self.entity_description.round_digits == 0:
-                return int(round(adjusted_value))
+                return round(adjusted_value)
             return round(adjusted_value, self.entity_description.round_digits)
 
         # Otherwise return raw value (e.g., strings for diagnostic fields)
